fixed compile warning

+++ ady_math.h
@@ -45,7 +45,7 @@ inline ivec2 IVec2(i32 x, i32 y)
 
 inline i32 Vec2To1DIndex(vec2 v, i32 width)
 {
-  return v.x + v.y * width;
+  return (i32)v.x + (i32)v.y * width;
 }
 
 // Vector 3 ----------------------------------------------------------------
